temporary remove settings
