+++ src/frompackage/input_parsing.jl
@@ -1,51 +1,3 @@
-function detect_custom_macro!(ex, macro_name::Symbol)
-    block = if Meta.isexpr(ex, :block)
-        ex
-    else
-        # We don't support custom macros for single statements
-        return nothing
-    end
-    idx = 0
-    for i ∈ eachindex(block.args)
-        arg = block.args[i]
-        if Meta.isexpr(arg, :macrocall) && arg.args[1] === macro_name
-            idx = i
-            break
-        end
-    end
-    idx != 0 || return nothing
-    # We extract the arg and delete it from the expr
-    arg = block.args[idx]
-    deleteat!(block.args, idx)
-    return arg
-end
-
-# This function checks for settings and eventually stores them in the package_dict
-function process_settings!(ex, dict)
-    settings_arg = detect_custom_macro!(ex, Symbol("@settings"))
-    !isnothing(settings_arg) || return ex
-    parse_settings(settings_arg, dict)
-    return ex
-end
-
-function parse_settings(ex, dict)
-    maybe_block_arg = ex.args[3]
-    setting_args = if Meta.isexpr(maybe_block_arg, :block)
-        maybe_block_arg.args
-    else
-        ex.args[3:end]
-    end
-    for arg in setting_args
-        arg isa LineNumberNode && continue
-        @assert Meta.isexpr(arg, :(=)) && length(arg.args) == 2 "Only `var = value` statements are allowed within the `@settings` block"
-        custom_settings = get!(Dict{Symbol,Any}, dict, "Custom Settings")
-        k, v = arg.args
-        @assert !(v isa Expr) "Only primitive values are allowed as values in the `@settings` block"
-        name = Settings.setting_name(k)
-        custom_settings[name] = v
-    end
-end
-
 # New stuff
 
 function should_exclude_using_names!(ex::Expr)
